requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
pymongo